test(synth): Handle upper coordinate space IDs in `RLE`/`Delta`.. logcial space (#1472)
Adds Rust-only synthetic fixtures for u32::MAX and u64::MAX feature ids
under the delta, rle, and delta_rle logical encodings; the u32
delta-encoded ones are decoded incorrectly by the C++ decoder today
(sign-extended to u64), which these fixtures pin down.

Both TS and CPP fail this.
TS does confuse `u32::MAX` being `-1`, so a signed/unsigned bug
CPP does assume that `u32::MAX` and `u64::MAX` fit the same space and
that `-1` is the same for both. It is not, so incorrect.

Maybe resolves
https://github.com/maplibre/maplibre-tile-spec/issues/1075

Hard to tell..

+++ cpp/src/mlt/decoder.cpp
@@ -79,7 +79,8 @@ Layer Decoder::Impl::parseBasicMVTEquivalent(BufferStream& tileData) {
             if (columnMetadata.getScalarType().hasLongID) {
                 integerDecoder.decodeIntStream<std::uint64_t>(tileData, denseIds, *idDataStreamMetadata);
             } else {
-                integerDecoder.decodeIntStream<std::uint32_t, std::uint64_t>(tileData, denseIds, *idDataStreamMetadata);
+                // Decode 32-bit ids at 32-bit width so delta math wraps there, then widen to id_t.
+                integerDecoder.decodeIntStream<std::uint32_t, std::uint32_t>(tileData, denseIds, *idDataStreamMetadata);
             }
 
             if (!idPresentBits.empty()) {
